(vt) Give Layer and ProtobufLayer their own files
